Code style cleanup
