+++ src/Workspaces/Core/Portable/Options/OptionSet.cs
@@ -12,7 +12,7 @@ using Microsoft.CodeAnalysis.Diagnostics;
 
 namespace Microsoft.CodeAnalysis.Options
 {
-    public abstract class OptionSet
+    public abstract partial class OptionSet
     {
         private const string NoLanguageSentinel = "\0";
         private static readonly ImmutableDictionary<string, AnalyzerConfigOptions> s_emptyAnalyzerConfigOptions =
@@ -78,36 +78,9 @@ namespace Microsoft.CodeAnalysis.Options
 
         internal abstract IEnumerable<OptionKey> GetChangedOptions(OptionSet optionSet);
 
-        internal virtual AnalyzerConfigOptions CreateAnalyzerConfigOptions(IOptionService optionService, string? language)
+        private protected virtual AnalyzerConfigOptions CreateAnalyzerConfigOptions(IOptionService optionService, string? language)
         {
             return new AnalyzerConfigOptionsImpl(this, optionService, language);
-        }
-
-        private sealed class AnalyzerConfigOptionsImpl : AnalyzerConfigOptions
-        {
-            private readonly OptionSet _optionSet;
-            private readonly IOptionService _optionService;
-            private readonly string? _language;
-
-            public AnalyzerConfigOptionsImpl(OptionSet optionSet, IOptionService optionService, string? language)
-            {
-                _optionSet = optionSet;
-                _optionService = optionService;
-                _language = language;
-            }
-
-            public override bool TryGetValue(string key, out string? value)
-            {
-                if (!_optionService.TryMapEditorConfigKeyToOption(key, _language, out var storageLocation, out var optionKey))
-                {
-                    value = null;
-                    return false;
-                }
-
-                var typedValue = _optionSet.GetOption(optionKey);
-                value = storageLocation.GetEditorConfigString(typedValue, _optionSet);
-                return true;
-            }
         }
     }
 }
